# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orion — Declarative Services Runtime API",
    "description": "Declarative services runtime platform.\n\n**Authentication.** The admin API (`/api/v1/admin/*`), the Prometheus endpoint (`/metrics`), and the trace read endpoints (`/api/v1/admin/traces*`) require an admin API key when `admin_auth.enabled` is true — the default in the shipped Helm chart and HA compose files. Operations that need it carry a `security` block; see the `admin_bearer` / `admin_api_key` schemes for how the key is presented and for the `admin_auth.header` config key that selects between them.\n\nThe data plane (`/api/v1/data/{channel}`) and the health probes are deliberately unauthenticated: channel-level access control is expressed through each channel's `validation_logic` and CORS configuration.\n\n**Errors.** Every non-2xx response uses the `ErrorResponse` envelope: `{\"error\": {\"code\", \"message\", \"request_id\"}}`, plus `details` on validation failures.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/channels/{id}": {
      "get": {
        "tags": [
          "Channels"
        ],
        "operationId": "get_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Channel details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ChannelResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "Channels"
        ],
        "operationId": "update_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Draft channel updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ChannelResponse"
                }
              }
            }
          },
          "400": {
            "description": "No draft version or invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Channels"
        ],
        "operationId": "delete_channel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Channel deleted"
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Channel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      }
    },
...
    "/api/v1/admin/traces": {
      "get": {
        "tags": [
          "Traces"
        ],
        "operationId": "list_traces",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by trace status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "description": "Filter by channel",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "description": "Filter by mode: sync, async",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (default 50, max 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Page offset",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Sort column: created_at (default), updated_at, status, channel, mode",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "description": "Sort direction: asc or desc (default)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of traces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEnvelope_TraceListItem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error response body matching Orion's\n`{\"error\": {\"code\": \"...\", \"message\": \"...\", \"request_id\": \"...\"}}` format.\n\nEvery 4xx/5xx from `OrionError` — and the 500 emitted by the panic-recovery\nlayer — uses this envelope.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "UpdateChannelRequest": {
        "type": "object",
        "properties": {
          "config": {},
          "consumer_group": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "methods": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "route_pattern": {
            "type": [
              "string",
              "null"
            ]
          },
          "topic": {
            "type": [
              "string",
              "null"
            ]
          },
          "transport_config": {},
          "workflow_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Alternative to `admin_bearer`, active when `admin_auth.header` names a header other than `Authorization`. The raw key is sent as that header's value with no `Bearer ` prefix. `X-API-Key` is the conventional choice and is what this document shows, but the header name is deployment-specific — substitute whatever `admin_auth.header` is set to. Exactly one of the two schemes is live at a time; they are listed as alternatives because this document is generated without knowledge of a deployment's config."
      },
      "admin_bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin API key presented as `Authorization: Bearer <key>` — the default. Enforced on `/api/v1/admin/*`, `/metrics`, and `/api/v1/admin/traces*` whenever `admin_auth.enabled` is true, which the shipped Helm chart and HA compose files set. Keys come from `admin_auth.api_keys`, either in plaintext or as `sha256:<64-hex>` digests. The data plane (`POST /api/v1/data/{channel}`) is not covered by this scheme."
      }
    }
  },
...
}